requirements.txt
flask
groq
python-dotenv
nltk
language-tool-python
setuptools
python-docx
gunicorn
openai==1.3.5
PyPDF2==3.0.1
requests==2.31.0
google-generativeai==0.3.2
reportlab==4.0.8
pytesseract==0.3.10
